requirements.txt
Django==3.2.25
djangorestframework==3.15.1
joblib==1.3.2
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-server==1.24.0
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
matplotlib==3.4.3
matplotlib-inline==0.1.6
numpy==1.17.5
opencv-contrib-python==3.4.2.17
opencv-python==4.5.4.58
opencv-python-headless==4.2.0.34
packaging==21.2
pandas==1.3.5
Pillow==8.4.0
pyspark==3.4.3
requests==2.31.0
scikit-learn==1.0.2
scipy==1.7.2
tqdm==4.66.1
